requirements.txt
django
djangorestframework
psycopg2_binary
pytest
pytest-django
factory_boy
pytest_factoryboy
drf-nested-routers